% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_consistent_filter.R
\name{condition_consistent_filter}
\alias{condition_consistent_filter}
\title{Condition consistency filter}
\usage{
condition_consistent_filter(x, rthr, cond)
}
\arguments{
\item{x}{the matrix to be filtered. It must have a \code{circ_id} column}

\item{rthr}{the minimum BJR count for a circRNA to be kept}

\item{cond}{the sample metadata table. Must have \code{sample_id} and
\code{condition} columns}
}
\value{
the matrix bearing circRNAs expressed in all samples of at least one
  condition
}
\description{
CircRNAs must be detected in all the samples of at least one condition
}
