% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_of_samples_filter.R
\name{fraction_the_samples_filter}
\alias{fraction_the_samples_filter}
\title{Fraction of samples filter}
\usage{
fraction_the_samples_filter(x, rthr = 0, frac = 0.5)
}
\arguments{
\item{x}{the matrix to be filtered. It must have a \code{circ_id} column}

\item{rthr}{the minimum BJR count for a circRNA to be kept}

\item{frac}{the fraction of the samples in which each circRNA must be
detected. Default is that each circRNA must be detected in al least half
the samples}
}
\value{
the matrix bearing circRNAs detected in at least a fraction of the
  samples
}
\description{
Each circRNA must be detected in at least the specified fraction of samples
}
