% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_methods.R
\name{get_combined_matrix}
\alias{get_combined_matrix}
\title{Combine different methods' expression estimates into one expression matrix}
\usage{
get_combined_matrix(x, filtfun = NULL, name_sep = ".", as_data_table = F)
}
\arguments{
\item{x}{the list of the methods' output to be combined or the path of the
CirComPara2's results}

\item{filtfun}{a function to filter the rows in the combined matrix}

\item{name_sep}{the character(s) used to concatenate SAMPLE and METHOD names}

\item{as_data_table}{if TRUE, return a data.table instead of a matrix}
}
\value{
a matrix of the expression values with samples in columns and
  features in the rows. Column names have the form SAMPLE.METHOD.
}
\description{
Combine different methods' expression estimates into one expression matrix
}
\examples{
\dontrun{mat <- get_combined_matrix(x = "/home/user/CirComPara2_analysis")}
}
