% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallest_group_filter.R
\name{smallest_group_filter}
\alias{smallest_group_filter}
\title{Smallest condition size filter}
\usage{
smallest_group_filter(x, rthr, cond)
}
\arguments{
\item{x}{the matrix to be filtered. It must have a \code{circ_id} column}

\item{rthr}{the minimum BJR count for a circRNA to be kept}

\item{cond}{the sample metadata table. Must have \code{sample_id} and
\code{condition} columns}
}
\value{
a matrix bearing circRNAs expressed in as many samples as the size of
  the smallest condition group
}
\description{
Each circRNA must be detected in at least as many samples as in the smallest
condition group
}
